skal ikke vise avbrutte eller feilregistrerte dokumenter i høyremenyen
