+++ src/frontend/Komponenter/Behandling/Høyremeny/Dokumentoversikt/LogiskeVedlegg.tsx
@@ -1,21 +1,21 @@
 import styled from 'styled-components';
 import * as React from 'react';
 import { Detail } from '@navikt/ds-react';
-import { LogiskVedlegg } from '../../../App/typer/dokumentliste';
+import { LogiskVedlegg } from '../../../../App/typer/dokument';
 
-const LogiskVedleggWrapper = styled.ul`
+const Container = styled.ul`
     list-style-type: circle;
 `;
 
 export const LogiskeVedlegg: React.FC<{ logiskeVedlegg: LogiskVedlegg[] | undefined }> = ({
     logiskeVedlegg,
 }) => (
-    <LogiskVedleggWrapper>
+    <Container>
         {logiskeVedlegg &&
             logiskeVedlegg.map((logiskVedlegg, index) => (
                 <li key={logiskVedlegg.tittel + index}>
                     <Detail>{logiskVedlegg.tittel}</Detail>
                 </li>
             ))}
-    </LogiskVedleggWrapper>
+    </Container>
 );
